requirements.txt
langchain
openai
fastapi
pydantic
python-dotenv
uvicorn
